+++ mod3rd/__init__.py
@@ -1,0 +1,13 @@
+"""
+
+this contains modules not supported by modcore directly
+
+"""
+
+print()
+print( "loading unsupported modules" )
+print()
+
+
+import mod3rd.admin_esp.wlan
+