+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - reworked fiber samples
 - FibreWatchdog timer exposed
 - custom HTML 404 handling
+- simple admin html app for wlan (under unsupported mod3rd folder)
+- wlan module rework, scan added, reconnect
 -